lesson-08/src/core/config_manager.py
"""
설정 관리 - 시스템 설정 로드 및 관리
"""

import yaml
from typing import Dict, Any, Optional
from pathlib import Path


class ConfigManager:
    """
    시스템 설정 관리자
    """
    
    def __init__(self, config_path: str = "config/config.yaml"):
        """
        설정 관리자 초기화
        
        Args:
            config_path: 설정 파일 경로
        """
        self.config_path = Path(config_path)
        self.config: Dict[str, Any] = {}
        
    def load_config(self) -> Dict[str, Any]:
        """
        설정 파일 로드
        
        Returns:
            설정 딕셔너리
        """
        # TODO: 구현 필요
        pass
    
    def save_config(self, config: Dict[str, Any]) -> bool:
        """
        설정 파일 저장
        
        Args:
            config: 저장할 설정
            
        Returns:
            성공 여부
        """
        # TODO: 구현 필요
        pass
    
    def validate_config(self, config: Dict[str, Any]) -> bool:
        """
        설정 검증
        
        Args:
            config: 검증할 설정
            
        Returns:
            검증 통과 여부
        """
        # TODO: 구현 필요
        pass

