lesson-08/requirements.txt
# CryptoAutoTrader 의존성 패키지

# 웹 프레임워크
Flask==2.3.2
flask-cors==4.0.0

# 데이터 처리
pandas==2.0.3
numpy==1.24.3

# API 클라이언트
requests==2.31.0
websocket-client==1.6.1

# 암호화폐 거래소
pyupbit==0.2.31

# 데이터베이스
SQLAlchemy==2.0.19

# 설정 파일
PyYAML==6.0.1

# 텔레그램 봇
python-telegram-bot==20.4

# 차트 및 시각화
matplotlib==3.7.2
plotly==5.15.0

# 기술적 지표
ta-lib==0.4.28
pandas-ta==0.3.14b0

# 유틸리티
python-dotenv==1.0.0
python-dateutil==2.8.2

# 테스트
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# 코드 품질
pylint==2.17.4
black==23.7.0
flake8==6.0.0

# 타입 체크
mypy==1.4.1

# 로깅
colorlog==6.7.0

